:sparkles: Add some bonus exercises

+++ approche-objet/src/main/java/banque/entites/Compte.java
@@ -3,9 +3,9 @@ package banque.entites;
 public class Compte {
 
     public String numero;
-    public Integer solde;
+    public double solde;
 
-    public Compte(String numero, Integer solde){
+    public Compte(String numero, double solde){
         this.numero = numero;
         this.solde = solde;
     }
